Changed to ajax-i
The menu on the left is 'ajaxi' in that the links don't get a complete
page refresh anymore, but rather - a click on a link loads the div
id="xboard-main" with the content of the view.

This means that on the controller (i.e. basics) - the layout to use
should be ajax (so empty)

Note that this creates a conflict with dynamic javascripts on views.
there are ways around it (mainly - extracting all js functions such as
the login_popup in team.ctp into a stand-alone js file and managing
events this way. However, there may be pages when this won't be
applicable and we'll need to resort to full GET.